% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_sample_metadata}
\alias{add_sample_metadata}
\title{Add Sample Metadata}
\usage{
add_sample_metadata(wells, bio_rep_re = "\\\\s\\\\d+$")
}
\arguments{
\item{wells}{A data frame containing wells information.}

\item{bio_rep_re}{A regular expression pattern used to remove the biological replicate number from the sample name. Defaults to `"\\s\\d+$"`.}
}
\value{
A modified data frame with additional columns: `sample_type`, `bio_rep`, and `tech_rep`.
}
\description{
Adds metadata to a wells data frame by extracting the sample type, biological replicate,
and calculating technical replicate numbers.
}
