% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{average_housekeeping}
\alias{average_housekeeping}
\title{Calculate Average Housekeeping Cq}
\usage{
average_housekeeping(wells, housekeeping_target)
}
\arguments{
\item{wells}{A data frame containing wells information.}

\item{housekeeping_target}{A character string specifying the housekeeping gene target.}
}
\value{
A data frame with columns `sample` and `mean_cq`, representing the average Cq value per sample.
}
\description{
Filters the wells for a specific housekeeping target and computes the average Cq value for each sample.
}
