% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_DCq}
\alias{calculate_DCq}
\title{Calculate Delta Cq (DCq)}
\usage{
calculate_DCq(wells, housekeeping_target)
}
\arguments{
\item{wells}{A data frame containing wells information.}

\item{housekeeping_target}{A character string specifying the housekeeping gene target.}
}
\value{
A data frame with an added column `DCq` representing the difference between the Cq and the mean housekeeping Cq.
}
\description{
Computes the Delta Cq (DCq) values by subtracting the mean housekeeping Cq from each sample's Cq.
}
