% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DCT.R
\name{calculate_DCT}
\alias{calculate_DCT}
\title{Caclulate Delta CT}
\usage{
calculate_DCT(df, cols)
}
\arguments{
\item{df}{hkg df}

\item{cols}{Column names to transfer to DCT df}
}
\value{
DCT df
}
\description{
Merges hkg with qpcr df, removes samples with bad hkg CT levels and calculates Delta CT
}
