% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_hkg.R
\name{calculate_mean_hkg}
\alias{calculate_mean_hkg}
\title{Calculate mean housekeeping genes}
\usage{
calculate_mean_hkg(df)
}
\arguments{
\item{df}{hkg df}
}
\value{
mean_hkg df
}
\description{
Calculate the average CT value over housekeeping genes
}
