% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_relative_expression}
\alias{calculate_relative_expression}
\title{Calculate Relative Expression}
\usage{
calculate_relative_expression(wells)
}
\arguments{
\item{wells}{A data frame containing wells information with a `DCq` column.}
}
\value{
A data frame with an additional column `rel_exp` representing the relative expression.
}
\description{
Computes the relative expression values using the formula 2^(-DCq).
}
