% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{cleanup}
\alias{cleanup}
\title{Remove NA sample wise as foundation of valid mean relative expression calculation}
\usage{
cleanup(df, hkg_)
}
\arguments{
\item{df}{initial data frame of expression data}

\item{hkg_}{character string of provided housekeeping genes}
}
\description{
Remove NA sample wise as foundation of valid mean relative expression calculation
}
\examples{
cleanup(df, khg_ = c("HKG"))
}
