% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{conflate}
\alias{conflate}
\title{Summarize expression data}
\usage{
conflate(df, do)
}
\arguments{
\item{df}{data frame of relative expression data}

\item{do}{boolean if to aggregate the input data (default: TRUE)}
}
\description{
Summarize expression data
}
\examples{
conflate(df)
}
