% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{control_mean}
\alias{control_mean}
\title{get mean expression values of all control groups}
\usage{
control_mean(df)
}
\arguments{
\item{df}{clean data frame of expression data}
}
\description{
get mean expression values of all control groups
}
\examples{
control_mean(df)
}
