% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{cut_in_half}
\alias{cut_in_half}
\title{Cut vector into half}
\usage{
cut_in_half(x, chunk_size = 2)
}
\arguments{
\item{x}{numeric vector}

\item{chunk_size}{number of elements per chunk (default: 2)}
}
\value{
vector half in length of input vector
}
\description{
Cut vector into half
}
\keyword{internal}
