% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{cycle_threshold}
\alias{cycle_threshold}
\title{The Ct value of real-time PCR}
\usage{
cycle_threshold(x, s_t)
}
\arguments{
\item{x}{a dataframe or matrix with two or more columns.
The name of first columns has to be "Cycle",and the names from the
second column to the last column are "Well".}

\item{s_t}{The single threshold of qPCR.}
}
\description{
The Ct value of real-time PCR
}
\examples{
## cycle_threshold(x, s_t)
}
