% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCt.R
\name{dct}
\alias{dct}
\title{Delta Ct calculation}
\usage{
dct(df)
}
\arguments{
\item{dataframe}{}
}
\value{
vector
}
\description{
Subtracts Ct of housekeeping gene from Ct of experimental target and returns values as a vector
}
\examples{
delta_Ct_values <- dCt(df)

}
