% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCt2.R
\name{dCt2}
\alias{dCt2}
\title{Delta Ct calculation}
\usage{
dCt2(df, endog_cntrl, cntrl_num = 1, excld_cntrl)
}
\arguments{
\item{df}{dataframe: most often the output from tidy_qPCR}

\item{endog_cntrl}{character: list of character names for endogenous control genes to be averaged}

\item{cntrl_num}{numeric: integer indicating number of endogenous controls to average}

\item{excld_cntrl}{character: list of endogenous control genes to exclude if not averaging between multiple}
}
\value{
data frame
}
\description{
Subtracts Ct of housekeeping gene from Ct of experimental target and returns values as a new column to input data frame
}
\examples{
qPCR_df <- df \%>\%
tidy_qPCR() \%>\%
dCt2()

}
