% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddCt.R
\name{ddCt}
\alias{ddCt}
\title{Delta Delta Ct calculation}
\usage{
ddCt(df, cntrl_cond)
}
\arguments{
\item{df}{data frame containing column with dCt values generated by dCt()}

\item{cntrl_cond}{the sample_name of the control condition in the experiment}
}
\value{
data frame
}
\description{
Subtracts the delta Ct values from a control condition from the delta Ct values of experimental condition
}
\examples{
ddCt_df <- ddCt(df)

}
