% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{delta_cq}
\alias{delta_cq}
\title{calculate delta cq values per comparison}
\usage{
delta_cq(df, contr_mean)
}
\arguments{
\item{df}{clean data frame of expression data}

\item{contr_mean}{object of control_mean function}
}
\description{
calculate delta cq values per comparison
}
\examples{
delta_cq(df, contr_mean)
}
