% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{detect_reference}
\alias{detect_reference}
\title{Detect variable of input data used as 'reference'}
\usage{
detect_reference(treatment, reference)
}
\arguments{
\item{treatment}{vector of unique character strings found in input data ('treatment' column)}

\item{reference}{vector of single reference value}
}
\value{
character string used as 'reference'
}
\description{
Detect variable of input data used as 'reference'
}
\keyword{internal}
