% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{drop_columns}
\alias{drop_columns}
\title{Drop columns based on character vector}
\usage{
drop_columns(df, cols = c("brep", "trep", "id"))
}
\arguments{
\item{df}{input data frame}

\item{cols}{character vector of columns to drop}
}
\value{
data frame with omitted columns
}
\description{
Drop columns based on character vector
}
\keyword{internal}
