% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{efficiency}
\alias{efficiency}
\title{Amplification efficiency}
\usage{
efficiency(x)
}
\arguments{
\item{x}{a dataframe or matrix.It must have the following columns:
Target, Sample, Ct, Log.Starting.Quantity}
}
\description{
Amplification efficiency
}
