% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{filter_NA}
\alias{filter_NA}
\title{Sanitize groups if NA in input data}
\usage{
filter_NA(df)
}
\arguments{
\item{df}{input data.frame containing NA}
}
\value{
data frame without NA values in cq column by removing sample specific values per group to ensure balanced number of observations
}
\description{
Sanitize groups if NA in input data
}
\keyword{internal}
