% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sample.R
\name{filter_sample}
\alias{filter_sample}
\title{Filter for a sample number}
\usage{
filter_sample(df, sample_name)
}
\arguments{
\item{df}{qpcr file}

\item{sample_name}{Name of sample to filter}
}
\value{
Tibble
}
\description{
Outputs a tibble with information about this specific sample
}
