% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{geometric_mean}
\alias{geometric_mean}
\title{Calculate geometric mean if multiple housekeeping genes are given}
\usage{
geometric_mean(values)
}
\arguments{
\item{values}{vector of numeric values}
}
\value{
geometric mean of numeric input vector
}
\description{
Calculate geometric mean if multiple housekeeping genes are given
}
\keyword{internal}
