% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{get_control_group}
\alias{get_control_group}
\title{select control group of data}
\usage{
get_control_group(df, control = "control")
}
\arguments{
\item{df}{data frame of provided genes}
}
\description{
select control group of data
}
\examples{
get_control_group(df)
}
