% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{get_E}
\alias{get_E}
\title{Calculate E value of standard samples}
\usage{
get_E(efficiency)
}
\arguments{
\item{efficiency}{numeric vector of efficiency values in percent (vector, 0-100)}
}
\value{
numeric vector of E values based on provided efficiency
}
\description{
Calculate E value of standard samples
}
\examples{
get_E(c(87,67,98,78))
}
