% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Gets your raw qPCR data from xls file}
\usage{
get_raw_data(
  file_name,
  ct_skip_rows = 7,
  ct_sheet = 2,
  amp_skip_rows = 7,
  amp_sheet = 5,
  meltcurve_skip_rows = 7,
  meltcurve_sheet = 9,
  amp_data = TRUE,
  meltcurve_data = TRUE
)
}
\arguments{
\item{file_name}{file name of your xls qPCR data}

\item{ct_skip_rows}{Number of rows to skip before header of Results table (default = 7)}

\item{ct_sheet}{Excel sheet number Results table is on with raw Ct results (default = 2)}

\item{amp_skip_rows}{Number of rows to skip before header of Amplification table (default = 7)}

\item{amp_sheet}{Excel sheet number Amplification data is on (default = 5)}

\item{meltcurve_skip_rows}{Number of rows to skip before header of Meltucurve table (default = 7)}

\item{meltcurve_sheet}{Excel sheet number Meltcurve data is on (default = 9)}

\item{amp_data}{Do you have amplification data to include (TRUE/FALSE)}

\item{meltcurve_data}{Do you have meltcurve data to include (TRUE/FALSE)}
}
\value{
A list of results tables as tibbles; table with raw Ct values, table with amplification values, table with meltcurve values
}
\description{
Gets your raw qPCR data from xls file
}
\examples{
my_data <- get_raw_data("extdata/test.xls")
}
