% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{get_reference_group}
\alias{get_reference_group}
\title{Select control group of data}
\usage{
get_reference_group(df)
}
\arguments{
\item{df}{data frame of provided genes}
}
\value{
data frame of selected reference group (treatment == reference)
}
\description{
Select control group of data
}
\keyword{internal}
