% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Take the geometric average}
\usage{
gm_mean(a, na.rm = TRUE)
}
\arguments{
\item{a}{vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
Geometric average
}
\description{
Calculates the geometric average over a vector
}
