% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_diff.R
\name{Group_diff}
\alias{Group_diff}
\title{Calculates differences in fold induction for two groups of samples
This function takes as input a data frame where the first n columns contain
categorical variables describing each sample (treatment, experiment, cell,...)
and the remaining columns containing expression values (as fold over mean)
for different genes (one gene per column). The function calculates the 4
difference in fold expression for each gene between two groups of samples
defined by the values in one of the columns describing the samples
@param df data matrix containing sample info in the first n columns and Ct
for measured genes (one gene per column)
@param nInfCol number of columns containing sample information. This columns
must precede those with Ct. The data frame must contain at least one column
containing sample information to group data.
@param colGroup number of column containing the variable to group data (only)
two categories are allowed.
@return a data frame containing the first n columns with sample information
followed by the differences in fold values for each gene
@export}
\usage{
Group_diff(df, nInfCol, colGroup)
}
\description{
Calculates differences in fold induction for two groups of samples
This function takes as input a data frame where the first n columns contain
categorical variables describing each sample (treatment, experiment, cell,...)
and the remaining columns containing expression values (as fold over mean)
for different genes (one gene per column). The function calculates the 4
difference in fold expression for each gene between two groups of samples
defined by the values in one of the columns describing the samples
@param df data matrix containing sample info in the first n columns and Ct
for measured genes (one gene per column)
@param nInfCol number of columns containing sample information. This columns
must precede those with Ct. The data frame must contain at least one column
containing sample information to group data.
@param colGroup number of column containing the variable to group data (only)
two categories are allowed.
@return a data frame containing the first n columns with sample information
followed by the differences in fold values for each gene
@export
}
