% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importqpcr.R
\name{importqpcr}
\alias{importqpcr}
\title{An import qpcr function.}
\usage{
importqpcr(workdir = ".")
}
\arguments{
\item{workdir}{Your working directory where to look for txt files. Defaults to current directory.}
}
\description{
This function will import raw exported txt file from BioRad machine (make sure to check all fields when exporting!) and return a table of values including all fields and NA.
}
\examples{
importqpcr()
}
