% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{in_range}
\alias{in_range}
\title{Check if value(s) is in specific range}
\usage{
in_range(x, min_, max_)
}
\arguments{
\item{x}{single value or vector to test}

\item{min_}{minimal allowed value}

\item{max_}{maximal allowed value}
}
\value{
boolean value of all values are in range
}
\description{
Check if value(s) is in specific range
}
\keyword{internal}
