% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_and_clean_qpcr.R
\name{join_and_clean_qpcr}
\alias{join_and_clean_qpcr}
\title{Merge and clean data}
\usage{
join_and_clean_qpcr(df = qpcr, df.a = ann_samples, df.b = ann_genes,
  by.a = "Sample", by.b = "Gene")
}
\arguments{
\item{df}{qpcr file}

\item{df.a}{Samples annotation file}

\item{df.b}{Genes annotation file}

\item{by.a}{column to join annotation samples}

\item{by.b}{column to join gene samples}
}
\value{
A tidy and clean dataframe
}
\description{
Merges data and annotation files together and changes the columns in the right types
}
