% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_double_meltcurves.R
\name{list_double_meltcurves}
\alias{list_double_meltcurves}
\title{Filter for primers with double meltcurve peaks}
\usage{
list_double_meltcurves(df, CT = "CT")
}
\arguments{
\item{df}{qpcr df}
}
\value{
Tibble
}
\description{
Selects all observations that show more than one melt curve peak
}
