% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw.R
\name{load_raw}
\alias{load_raw}
\title{Import and Process Initial Raw Excel Data}
\usage{
load_raw(path, sheet, cell_range)
}
\value{
data frame
}
\description{
Takes QuantStudio output excel file and loads into df parsing CT as numeric
}
\examples{
load_raw()

}
