% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{make_groups}
\alias{make_groups}
\title{Create chunks of groups for calculation}
\usage{
make_groups(df, hkg, groups = NULL)
}
\arguments{
\item{df}{data frame of expression data}

\item{hkg}{character string of provided housekeeping gene(s)}

\item{groups}{character vector of requested groups to compare (default: NULL)}
}
\value{
a list of all target x housekeeping x group combinations
}
\description{
Create chunks of groups for calculation
}
\examples{
make_groups(df, hkg = "HKG", groups = c("plate", "diet", "timepoint"))
}
