% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_superplot}
\alias{make_superplot}
\title{Create a Superplot}
\usage{
make_superplot(wells, target, plot_order)
}
\arguments{
\item{wells}{A data frame containing wells information.}

\item{target}{A character string specifying the target gene to plot.}

\item{plot_order}{An optional character vector specifying the order of sample types. If provided, sample types are
converted to a factor with the specified levels.}
}
\value{
A ggplot object representing the superplot.
}
\description{
Generates a superplot that displays summary statistics (mean and standard error) as well as individual data points
for a specified target.
}
