% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeplots.R
\name{makeplots}
\alias{makeplots}
\title{A makeplots function
This function will make barplots with fold changes for your profiled genes.}
\usage{
makeplots(tab_cl, norm_gene = "Actin", cal = "mock", cell_type = "HEK293")
}
\arguments{
\item{tab_cl}{A table of your values in form "Sample, Target, Cq"}

\item{norm_gene}{A gene to which to normalize Ct values, normally housekeeping. Defaults to "Actin".}

\item{cal}{Your calibrator sample (control sample). In the plot of relative fold changes it will be set to 1. Defaults to "mock".}

\item{cell_type}{What cells do you do qpcr from? Needed for the title of your plot. Defaults to "HEK293".}
}
\description{
A makeplots function
This function will make barplots with fold changes for your profiled genes.
}
\examples{
makeplots()
}
