% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{mean_relative_expression}
\alias{mean_relative_expression}
\title{calculate mean relative expression}
\usage{
mean_relative_expression(df)
}
\arguments{
\item{df}{data frame of treatment and corresponding ratio values}
}
\description{
calculate mean relative expression
}
\examples{
mean_relative_expression(df)
}
