% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2ddCt.R
\name{nrmlzd_expr}
\alias{nrmlzd_expr}
\title{Calculation for 2^(-ddCt); the normalized expression value}
\usage{
nrmlzd_expr(df)
}
\arguments{
\item{df}{data frame containing column with dCt values generated by dCt()}

\item{cntrl_cond}{the sample_name of the control condition in the experiment}
}
\value{
data frame
}
\description{
Generates a normalized expression value in reference to a control condition based on ddCt
}
\examples{
final_df<- nrmlzd_expr(df)

}
