% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{outlier_method}
\alias{outlier_method}
\title{Check for outliers in expression values}
\usage{
outlier_method(x, method)
}
\arguments{
\item{x}{numeric vector of expression data (cq column)}

\item{method}{method to apply to find outliers of numeric vector}
}
\value{
boolean vector of numeric input data if outlier were detected by choosed method
}
\description{
Check for outliers in expression values
}
\examples{
outlier_method(x)
}
