% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot barplot}
\usage{
plot_bar(df, x.var, y.var, fill.var, ncol = 6, nrow = 4)
}
\arguments{
\item{df}{endog df}

\item{x.var}{Variable to plot on x-axis}

\item{y.var}{Variable to plot on y-axis}

\item{fill.var}{Variable to plot colours}

\item{ncol}{Amount of columns}

\item{nrow}{Amount of rows}
}
\value{
Bar plot
}
\description{
Plots a barplot for all genes
}
