% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar_per_gene}
\alias{plot_bar_per_gene}
\title{Plots a barplot per gene}
\usage{
plot_bar_per_gene(df, gene, x.var, y.var, fill.var)
}
\arguments{
\item{df}{endog df}

\item{gene}{Gene to plot}

\item{x.var}{Variable to plot on x-axis}

\item{y.var}{Variable to plot on y-axis}

\item{fill.var}{Variable to plot colours}
}
\description{
Plots a barplot per gene
}
