% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_data}
\alias{plot_data}
\title{Plot qPCR Data}
\usage{
plot_data(wells, target, plot_order = NULL)
}
\arguments{
\item{wells}{A processed data frame containing wells information.}

\item{target}{A character string specifying the target gene to plot.}

\item{plot_order}{An optional character vector specifying the order of sample types.}
}
\value{
A ggplot object representing the superplot.
}
\description{
Creates a superplot for a specified target using preprocessed data.
}
