% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pointrange.R
\name{plot_pointrange}
\alias{plot_pointrange}
\title{Plot pointrange}
\usage{
plot_pointrange(df, x.var = "Sample", y.var = "DCt", col.var, ncol = 6,
  nrow = 4)
}
\arguments{
\item{df}{DCT df}

\item{x.var}{Variable to plot on x-axis}

\item{y.var}{Variable to plot on y-axis}

\item{col.var}{Variable to plot colours}

\item{ncol}{Amount of columns}

\item{nrow}{Amount of rows}
}
\value{
Pointrange plot
}
\description{
Plots pointrange for all genes
}
