% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pointrange.R
\name{plot_pointrange_per_gene}
\alias{plot_pointrange_per_gene}
\title{Plots a pointrange per gene}
\usage{
plot_pointrange_per_gene(df, gene, x.var = "Sample", y.var = "DCt", col.var)
}
\arguments{
\item{df}{DCT df}

\item{gene}{Gene to plot}

\item{x.var}{Variable to plot on x-axis}

\item{y.var}{Variable to plot on y-axis}

\item{col.var}{Variable to plot colours}
}
\description{
Plots a pointrange per gene
}
