% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter_per_gene}
\alias{plot_scatter_per_gene}
\title{Plots a scatterplot per gene}
\usage{
plot_scatter_per_gene(df, gene, x.var, y.var, col.var = FALSE, theme = 14)
}
\arguments{
\item{df}{DCT df}

\item{gene}{Gene to plot}

\item{x.var}{Variable to plot on x-axis}

\item{y.var}{Variable to plot on y-axis}

\item{col.var}{Variable to plot colours}
}
\description{
Plots a scatterplot per gene
}
