% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{positive}
\alias{positive}
\title{Positive calling}
\usage{
positive(x, n, negative, method = "POR", p = 0.1)
}
\arguments{
\item{x}{a dataframe or matrix with two or more columns.
The name of first columns has to be "Cycle",and the names from the
second column to the last column are "Well".}

\item{n}{End cycles of every well}

\item{negative}{one or more wells which are negative control sample}

\item{method}{The calculation mode of tolerance.}

\item{p}{Percent of range or absolute RFU}
}
\description{
Positive calling
}
\details{
If method is POR(the default), the range of p is from 0.1(the default) to 0.9.
If method is RFU, p is a absolute RFU.
}
\examples{
## neg <- c("w1", "w2")
## positive(x, 5, neg)
## method is RFU
## positive(x, 5, neg, "RFU", 2500)
}
