% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{prepare}
\alias{prepare}
\title{Prepare input data}
\usage{
prepare(df)
}
\arguments{
\item{df}{input data.frame of qPCR data}
}
\value{
clean data frame of input data (type conversion, adding default values, calculate E values of efficiency)
}
\description{
Prepare input data
}
\keyword{internal}
