% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess qPCR Data}
\usage{
preprocess_data(file_path, housekeeping_target = "GAPDH", WT_string = "WT")
}
\arguments{
\item{file_path}{A character string specifying the path to the input data file.}

\item{housekeeping_target}{A character string specifying the housekeeping gene target. Defaults to `"GAPDH"`.}

\item{WT_string}{A character string used to identify wild-type samples. Defaults to `"WT"`.}
}
\value{
A processed data frame with computed columns including `DCq`, `rel_exp`, and `rel_exp_norm`.
}
\description{
Reads and preprocesses qPCR data from a file. The function cleans column names,
filters out omitted data, calculates Delta Cq (DCq), computes relative expression,
and normalizes the expression to the wild-type (WT) samples.
}
