% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mean_hkg.R
\name{qc_mean_hkg}
\alias{qc_mean_hkg}
\title{Plot QC graphs}
\usage{
qc_mean_hkg(df)
}
\arguments{
\item{df}{hkg df}
}
\value{
QC plots and mean_hkg df
}
\description{
Plot quality control graphs on average of housekeeping genes
}
