% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_tech_rep.R
\name{qc_tech_rep}
\alias{qc_tech_rep}
\title{Check quality technical replicates}
\usage{
qc_tech_rep(df, threshold = 0.4, col_target = "Target",
  col_sample = "Sample")
}
\arguments{
\item{df}{qpcr dataframe without annotations}
}
\value{
print of df with avg_CT, avg_Tm1 and avg_Tm2
}
\description{
Filters out technical replicates with too high stdev
}
