\name{qpcR.news}
\alias{qpcR.news}      

\title{Display news and changes of qpcR package versions}

\description{
Displays the latest changes (new functions, bug fixes etc.) of the different package versions in a text window. 
}

\usage{
qpcR.news(...)
}

\arguments{
  \item{...}{arguments to be passed to \code{\link{file.show}}. Usually needs no entry.}
}

\value{
None.
}

\author{
Andrej-Nikolai Spiess
}

\examples{
qpcR.news()    
}

\keyword{utilities}
\keyword{documentation}
