% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{qpcrviia7_theme_opts}
\alias{qpcrviia7_theme_opts}
\title{Theme for plotting}
\usage{
qpcrviia7_theme_opts()
}
\arguments{
\item{path}{Path to an excel file}
}
\value{
Theme command ggplot
}
\description{
Adds theme layer to ggplot object
}
