% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{ratio}
\alias{ratio}
\title{Calculate efficiency based ratio of delta cq value}
\usage{
ratio(df)
}
\arguments{
\item{df}{data frame to add ratio column to}
}
\description{
Calculate efficiency based ratio of delta cq value
}
\examples{
ratio(df)
}
