% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{ratio_by_mean_ratio}
\alias{ratio_by_mean_ratio}
\title{calculate the ratio compared to the mean ratio per gene}
\usage{
ratio_by_mean_ratio(d_cq, e_val, hkg_, treatm)
}
\arguments{
\item{d_cq}{object of delta_cq}

\item{e_val}{named list of efficiency values}

\item{hkg_}{character vector of housekeeping genes}

\item{treatm}{character vector of treatments}
}
\description{
calculate the ratio compared to the mean ratio per gene
}
\examples{
ratio_by_mean_ratio(d_cq, e_val, hkg_, treatm)
}
