% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annotation.R
\name{read_annotation}
\alias{read_annotation}
\title{Read annotation files}
\usage{
read_annotation(x, col_types = "guess")
}
\arguments{
\item{x}{An excel document}
}
\value{
A dataframe
}
\description{
Load annotation file into R
}
