% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr.R
\name{read_qpcr}
\alias{read_qpcr}
\title{Read qpcr file}
\usage{
read_qpcr(x, sheet = 3, skip = 35, melt_curves = TRUE)
}
\arguments{
\item{x}{An excel document}

\item{sheet}{Excel sheet to import}

\item{skip}{Rows to skip of excel sheet}

\item{melt_curves}{Boolean to integrate melt curve analysis}
}
\value{
A dataframe
}
\description{
Loads raw file with CT values from Viia7
}
