% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad_meltcurves.R
\name{remove_bad_meltcurves}
\alias{remove_bad_meltcurves}
\title{Remove observations with bad meltcurves}
\usage{
remove_bad_meltcurves()
}
\value{
df
}
\description{
Changes CT values of observations with more than one meltcurve to NA
}
