% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{Vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
vector without outliers
}
\description{
Change value of outliers to NA
}
