% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{sanitize}
\alias{sanitize}
\title{Remove NA sample wise as foundation of valid mean relative expression calculation}
\usage{
sanitize(list_of_groups, hkg_)
}
\arguments{
\item{list_of_groups}{output of 'make_groups' containing a list of groups to compare expression}

\item{hkg_}{character string of provided housekeeping genes}
}
\description{
Remove NA sample wise as foundation of valid mean relative expression calculation
}
\examples{
sanitize(list_of_groups, khg_ = c("HKG"))
}
