% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_CT_min_max.R
\name{set_CT_min_max}
\alias{set_CT_min_max}
\title{Set minimum and maximum CT value}
\usage{
set_CT_min_max(min = 10, max = 40)
}
\arguments{
\item{min}{Minimum CT value}

\item{max}{Maximum CT value}
}
\value{
qpcr dataframe
}
\description{
Changes out of range CT values to minimum or maximum number
}
