% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{set_efficiency}
\alias{set_efficiency}
\title{set efficiency of genes}
\usage{
set_efficiency(genes, efficiency, default = 100)
}
\arguments{
\item{genes}{character vector of gene names}

\item{efficiency}{named list of percentual efficiency of genes}

\item{default}{efficiency if not specified}
}
\description{
set efficiency of genes
}
\examples{
set_efficiency(c("gene1", "gene2"), list("gene1" = 98, "gene2" = 95))
}
