% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_min_max_CT.R
\name{set_min_max_CT}
\alias{set_min_max_CT}
\title{Set minimum and maximum CT value}
\usage{
set_min_max_CT(df, min = 10, max = 40)
}
\arguments{
\item{df}{Dataframe with raw CT values and column specifiying housekeeping genes}

\item{min}{Minimum CT value}

\item{max}{Maximum CT value}
}
\value{
qpcr dataframe
}
\description{
Changes out of range CT values to minimum or maximum number
}
