% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_CT_min_max.R
\name{set_min_max_CT_endog}
\alias{set_min_max_CT_endog}
\title{Set minimum and maximum CT value}
\usage{
set_min_max_CT_endog(min = 10, max = 40)
}
\arguments{
\item{min}{Minimum CT value}

\item{max}{Maximum CT value}

\item{df}{Dataframe with raw CT values and column specifiying housekeeping genes}
}
\value{
qpcr dataframe
}
\description{
Changes out of range CT values to minimum or maximum number
}
