% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_genes.R
\name{split_genes}
\alias{split_genes}
\title{Split genes}
\usage{
split_genes(df)
}
\arguments{
\item{df}{qpcr df}
}
\value{
endog Dataframe with only CT values from endogenous genes

hkg Dataframe with only CT values from housekeeping genes
}
\description{
Split data into endogenous genes and housekeeping genes
}
