% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{stat_t_test}
\alias{stat_t_test}
\title{Perform t-test}
\usage{
stat_t_test(df = endog, gene, group)
}
\arguments{
\item{df}{endog dataframe}

\item{gene}{Gene to perform t-test on}

\item{group}{Group variable between which the test should be performed}
}
\value{
p value
}
\description{
Performs t-test on Delta CT values of selected group and gene
}
