% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{threshold}
\alias{threshold}
\title{Single threshold of real-time PCR}
\usage{
threshold(x)
}
\arguments{
\item{x}{a dataframe or matrix with two or more columns.
The name of first columns has to be "Cycle",and the names from the
second column to the last column are "Well".}
}
\description{
Single threshold of real-time PCR
}
\examples{
## threshold(x)
}
