% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_qPCR.R
\name{tidy_qPCR}
\alias{tidy_qPCR}
\title{Initial tidying of raw qPCR data from QuantStudio}
\usage{
tidy_qPCR(df_raw, group_size, ref_tgt, exp_tgt)
}
\arguments{
\item{df_raw}{raw excel output from QuantStudio instrument}

\item{group_size}{numeric: number of rows that have a common target gene}

\item{ref_tgt}{character: name identifier of housekeeping/reference gene}

\item{exp_tgt}{character: name identifier of experimental target gene}
}
\value{
data frame
}
\description{
Applies R friendly names to variables, creates grouping index based on number of samples per target
}
\examples{
tidy_df <- tidy_qPCR()

}
