% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_qPCR2.R
\name{tidy_qPCR2}
\alias{tidy_qPCR2}
\title{Initial tidying of raw qPCR data from QuantStudio}
\usage{
tidy_qPCR2(df_raw, ref_tgt, exp_tgt)
}
\arguments{
\item{df_raw}{raw excel output from QuantStudio instrument}

\item{ref_tgt}{character: name identifier of housekeeping/reference gene(s)}

\item{exp_tgt}{character: name identifier of experimental target gene(s)}
}
\value{
data frame
}
\description{
Applies R friendly names to variables, tags endogenous controls ("house") or experimental targets by user input
}
\examples{
tidy_df <- tidy_qPCR()

}
