% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{comp_gene_pair}
\alias{comp_gene_pair}
\title{Get all combinations of target and housekeeping genes as list}
\usage{
comp_gene_pair(df, hkg)
}
\arguments{
\item{df}{data frame of every group}

\item{hkg}{character string of housekeeping gene(s)}
}
\value{
a list of all target x housekeeping gene combinations
}
\description{
Get all combinations of target and housekeeping genes as list
}
\keyword{internal}
