% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCt.R
\name{dCt}
\alias{dCt}
\title{Delta Ct calculation}
\usage{
dCt(df, group_size)
}
\arguments{
\item{df}{dataframe: most often the output from tidy_qPCR}

\item{group_size}{numeric: number of rows that have a common target gene}
}
\value{
data frame
}
\description{
Subtracts Ct of housekeeping gene from Ct of experimental target and returns values as a new column to input data frame
}
\examples{
qPCR_df <- dCt(df)

}
