% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ddct}
\alias{ddct}
\title{Comparative Ct method}
\usage{
ddct(x, goi, ref, ctrl)
}
\arguments{
\item{x}{a dataframe or matrix.It must have the following columns:
Target, Sample, Ct}

\item{goi}{gene of interest}

\item{ref}{reference gene}

\item{ctrl}{control sample of relative quantification in qPCR}
}
\description{
Comparative Ct method
}
\examples{
## ddct(x, goi, ref, ctrl)
}
