% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{delta_cq}
\alias{delta_cq}
\title{Calculate delta cq values per comparison}
\usage{
delta_cq(df)
}
\arguments{
\item{df}{clean data of given group}
}
\value{
a data frame with added columns 'ref_mean' and 'delta_cq' as prerequisite for relative expression calculation
}
\description{
Calculate delta cq values per comparison
}
\examples{
delta_cq(df)
}
