% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{filter_NA}
\alias{filter_NA}
\title{Sanitize groups if NA in input data}
\usage{
filter_NA(df)
}
\arguments{
\item{df}{input data.frame containing NA}
}
\description{
Sanitize groups if NA in input data
}
\examples{
sanitize(df)
}
