% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{filter_outlier}
\alias{filter_outlier}
\title{Filter outlier of expression values, gene and treatment specific}
\usage{
filter_outlier(df, method, do)
}
\arguments{
\item{df}{data frame of expression data}

\item{method}{method to apply to find outliers of numeric vector}

\item{do}{boolean if expression values should be filtered by outliers}
}
\value{
a data frame with removed outlier values based on a selected method ('interquartile', 'z-score', 'hampel')
}
\description{
Filter outlier of expression values, gene and treatment specific
}
\examples{
filter_outlier(df, method = "interquartile", do = TRUE)
}
