% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{pair_wise}
\alias{pair_wise}
\title{Apply calculation for every hkg vs. target pair}
\usage{
pair_wise(pair, hkg)
}
\arguments{
\item{pair}{named list of pairs (data)}

\item{hkg}{character vector of housekeeping genes}
}
\description{
Apply calculation for every hkg vs. target pair
}
\examples{
pair_wise(pair, e_values, hkg)
}
