% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{pair_wise}
\alias{pair_wise}
\title{Apply calculation for every HKG vs. target pair}
\usage{
pair_wise(pair, hkg)
}
\arguments{
\item{pair}{named list of pairs (data)}

\item{hkg}{character vector of housekeeping genes}
}
\value{
a lists of calculated ratios (data frame) for every housekeeping vs. target gene pair
}
\description{
Apply calculation for every HKG vs. target pair
}
\examples{
pair_wise(pair, hkg)
}
