% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot scatterplot}
\usage{
plot_scatter(df, x.var, y.var, col.var, ncol = 6, nrow = 4, theme = 8)
}
\arguments{
\item{df}{DCT df}

\item{x.var}{Variable to plot on x-axis}

\item{y.var}{Variable to plot on y-axis}

\item{col.var}{Variable to plot colours}

\item{ncol}{Amount of columns}

\item{nrow}{Amount of rows}
}
\value{
Scatter plot
}
\description{
Plots a scatterplot for all genes
}
