% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcR.r
\name{qpcR}
\alias{qpcR}
\title{main function to perform calculation of mean relative expression of qpcr data}
\usage{
qpcR(df, hkg = NULL, efficiency = NULL)
}
\arguments{
\item{df}{data frame of qpcr data}

\item{hkg}{character vector of housekeeping genes}

\item{efficiency}{named list of efficiency values}
}
\description{
main function to perform calculation of mean relative expression of qpcr data
}
\examples{
qpcr(df, hkg = c("HKG"))
}
