% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcR.r
\name{qpcR}
\alias{qpcR}
\title{Main function to perform calculation of relative expression values of qpcr data}
\usage{
qpcR(
  df,
  hkg = NULL,
  reference = NULL,
  groups = NULL,
  aggregate = TRUE,
  outlier = TRUE,
  outlier.method = "interquartile"
)
}
\arguments{
\item{df}{data frame of qpcr data}

\item{hkg}{character vector of housekeeping genes}

\item{reference}{character vector of reference sample (default = control)}

\item{groups}{character vector of conditions to group on}

\item{aggregate}{boolean if output should be aggregated by groups or not (reference will equal 1, default = TRUE)}

\item{outlier}{boolean if outlier should be detected and removed before calculation of relative expression values (default = TRUE)}

\item{outlier.method}{character string to choose the method to remove outliers (default: interquartile)}
}
\value{
data frame of relative expression values compared by provided groups and housekeeping genes
}
\description{
Main function to perform calculation of relative expression values of qpcr data
}
\examples{
qpcr(df, hkg = c("HKG"))
}
