% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.r
\name{ratio_by_mean_ratio}
\alias{ratio_by_mean_ratio}
\title{Calculate the ratio compared to the mean ratio per gene}
\usage{
ratio_by_mean_ratio(df, hkg)
}
\arguments{
\item{df}{data frame of requested groups}

\item{hkg}{character vector of housekeeping genes}
}
\value{
a data frame with calculted ratio of a gene compared to the mean ratio
}
\description{
Calculate the ratio compared to the mean ratio per gene
}
\examples{
ratio_by_mean_ratio(df, hkg)
}
