% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr.R
\name{read_qpcr}
\alias{read_qpcr}
\title{Read qpcr file}
\usage{
read_qpcr(x, sheet = 3, skip = 35)
}
\arguments{
\item{x}{An excel document}
}
\value{
A dataframe
}
\description{
Loads raw file with CT values from Viia7
}
