% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_blanc.R, R/remove_sample.R
\name{remove_sample}
\alias{remove_sample}
\alias{remove_sample}
\title{Remove blanc}
\usage{
remove_sample(df, sample_name)

remove_sample(df, sample_name)
}
\arguments{
\item{df}{qpcr file}

\item{sample_name}{Name of sample to remove}

\item{df}{qpcr file}
}
\value{
qpcr dataframe

qpcr dataframe
}
\description{
Changes CT values of a sample to NA

Changes CT values of a sample to NA
}
